Adding a new 'ref' subdirectory under isl/model for ISLJ/Java/MASON Reference Model implementation.
Also adding the ConvectionDispersion class which is the Java implementation of respective C++ implementation under src/refModel in the ISL/C/C++/ObjC/Swarm Reference Model.  I have started on the some of the other Reference Model classes but I'm not checking them in at this point because they need a bit more work.

The ConvectionDispersion class has a unit test class; see test/isl/model/ref/ConvectionDispersionTest.  My unit testing has shown that the while the Java and C++ produce very similar numbers in _absolute_ terms, on a _relative_ scale there is a huge error.  For example, for the first integrand calculation, both implementations are producing similar numbers that are on the order of 10^-45.  However when I calculate the number of representable double precision values between the C++ and Java outputs, they are off by hundreds of thousands.  As the integral calculations continue, the numbers look close but relatively speaking the error continues to grow, e.g. they are off by millions of possible/representable double precision values.

Most of the relative error between the Java and C++ Conventioal Convection Dispersion equation values seems to be attributed to the Bessel function implementation.  The C++ model is using the 'j0' function from the core C++ library and for the Java model we are using the Bessel class from the Colt 1.2.0 library (ie the CERN implementation, see the jar file in the 'lib' directory).

Anyway, I'm not planning to dig further into the relative error problem at the moment; instead I want to concentrate in finishing up the Extended Convection Dispersion calculations and the remainder of the Reference Model implementation.  When all that is complete, then I'll email my testing results to Glen so we can discuss the issue.

Directory level or SVN Property changes:
   * Set svn:ignore property to ignore '*.out' and '*.log' files in this directory and subdirectories.
   * Set svn:ignore property to ignore '.commit.txt' file in this directory and subdirectories.  The '.commit.txt' file is a scratch file that I'm using to compose commit messages and I'll likely want to keep the file laying around if I working on the same code and making successive checkins.

Ken