Major structural modifications.
My understanding of VTK is much better now and I have rewritten/replaced/removed much of the code. The remaining code is cleaner and more extensible. Further cleanup is underway.

Wiggling is not enabled in this update; instead, the cells are able to be rotated with the mouse.

Use vtkInput.txt as the input file.

V-Cell surface membranes are displayed as a sets of triangles. Some faces have holes and this needs to be fixed.